+++ R/generate_documentation.R
@@ -1,10 +1,10 @@
-#' @title A helper function for the UI part of the Shiny apps
+#' @title A helper function which processes and displays the documentation part for each app
 #'
 #' @description This function take the documentation provided as html file
 #' and extracts sections
 #' to generate the tabs with content for each Shiny app.
 #' This is a helper function and only useful for this package.
-#' @param docfilename of html file containing the documentation
+#' @param docfilename full path and name to html file containing the documentation
 #' @return tablist A list of tabs for display in a Shiny UI.
 #' @details This function is called by the Shiny UIs to populate the documentation and information tabs.
 #' @author Andreas Handel