+++ R/generate_ggplot.R
@@ -1,4 +1,4 @@
-#' @title A helper function that takes result from the simulators and produces plots
+#' @title A helper function that takes simulation results and produces ggplot plots
 #'
 #' @description This function generates plots to be displayed in the Shiny UI.
 #' This is a helper function. This function processes results returned from the simulation, supplied as a list.
@@ -24,22 +24,29 @@
 #'    optional: legendtitle - Legend title, if NULL/not supplied, default is used \cr
 #'    optional: legendlocation - if "left" is specified, top left. Otherwise top. \cr
 #'    optional: linesize - Width of line, numeric, i.e. 1.5, 2, etc. set to 1.5 if not supplied. \cr
+#'    optional: pallette - overwrite plot colors by providing a vector of color names or hex numbers to be used for the plot. \cr
 #'    optional: title - A title for each plot. \cr
 #'    optional: for multiple plots, specify res[[1]]$ncols to define number of columns \cr
 #'
 #' @return A ggplot plot structure for display in a Shiny UI.
-#' @details This function is called by the Shiny server to produce plots returned to the Shiny UI.
-#' Create plots run the simulation with default parameters just call the function:
-#' result <- simulate_basicbacteria()
-#' plot <- generate_ggplot(result)
+#' @details This function can be called to produce plots, i.e. those displayed for each app.
+#' The input needed by this function is produced by either calling the run_model() function (as done when going through the UI)
+#' or manually transforming the output from a simulate_ function into the correct list structure explained below.
 #' @rawNamespace import(ggplot2, except = last_plot)
 #' @importFrom stats reshape
 #' @importFrom gridExtra grid.arrange
+#' @importFrom rlang .data
 #' @author Andreas Handel
 #' @export
 
 generate_ggplot <- function(res)
 {
+
+    # change ggplot color palette to color-blind friendly
+    # http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/#a-colorblind-friendly-palette
+    # I added more colors at the end to have 12, enough for all simulations
+    # the ones I added are likely not color-blind friendly but rarely used in the app
+    cbfpalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7", "#00523B","#D5C711","#0019B2","#cc0000")
 
     #nplots contains the number of plots to be produced.
     nplots = length(res) #length of list
@@ -82,21 +89,21 @@ generate_ggplot <- function(res)
       {
         #using basic reshape function to reformat data
         dat = stats::reshape(rawdat, varying = colnames(rawdat)[-1], v.names = 'yvals', timevar = "varnames", times = colnames(rawdat)[-1], direction = 'long', new.row.names = NULL)
-		dat$id <- NULL
+		    dat$id <- NULL
       }
 
       #code variable names as factor and level them so they show up right in plot - factor is needed for plotting and text
       mylevels = unique(dat$varnames)
-      dat$varnames = factor(dat$varnames, levels = mylevels)
+      dat$varnames = factor(dat$varnames, levels = mylevels, ordered = TRUE)
 
       #see if user/calling function supplied x- and y-axis transformation information
       xscaletrans <- ifelse(is.null(resnow$xscale), 'identity',resnow$xscale)
       yscaletrans <- ifelse(is.null(resnow$yscale), 'identity',resnow$yscale)
-      
+
       #lower and upper bounds for plots, these are used if none are provided by calling function
-      lb = 1e-10;
-      ub = 1e20;
-      
+      lb = 1e-10
+      ub = 1e20
+
       #if we want a plot on log scale, set any value in the data at or below 0 to some small number
       if (xscaletrans !='identity') {dat$xvals[dat$xvals<=0]=lb}
       if (yscaletrans !='identity') {dat$yvals[dat$yvals<=0]=lb}
@@ -112,48 +119,50 @@ generate_ggplot <- function(res)
       #set line size as given by app or to 1.5 by default
       linesize = ifelse(is.null(resnow$linesize), 1.5, resnow$linesize)
 
-       #if the IDvar variable exists, use it for further stratification, otherwise just stratify on varnames
-	  if (is.null(dat$IDvar))
-      {
-        p1 = ggplot2::ggplot(dat, ggplot2::aes(x = xvals, y = yvals, color = varnames, linetype = varnames, shape = varnames) )
-      }
+
+      #if the IDvar variable exists, use it for further stratification, otherwise just stratify on varnames
+	    #the unusual notation for the aes settings is needed for use inside a package
+      #see here: https://ggplot2.tidyverse.org/dev/articles/ggplot2-in-packages.html
+      if (is.null(dat$IDvar))
+	    {
+	      p1 = ggplot2::ggplot(dat, ggplot2::aes(x = .data$xvals) )
+	    }
       else
       {
-        p1 = ggplot2::ggplot(dat, ggplot2::aes(x = xvals, y = yvals, color = varnames, linetype = varnames, group = IDvar) )
+        p1 = ggplot2::ggplot(dat, ggplot2::aes(x = .data$xvals, group = .data$IDvar) )
       }
 
       ###choose between different types of plots
       if (plottype == 'Scatterplot')
       {
-        p2 = p1 + ggplot2::geom_point( size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_point(data = dat, aes( y = .data$yvals, color = .data$varnames, shape = .data$varnames), size = linesize, na.rm=TRUE)
       }
       if (plottype == 'Boxplot')
       {
-        p2 = p1 + ggplot2::geom_boxplot(size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_boxplot(data = dat, aes( y = .data$yvals, color = .data$varnames), size = linesize, na.rm=TRUE)
       }
       if (plottype == 'Lineplot')
       {
-        p2 = p1 + ggplot2::geom_line(size = linesize, na.rm=TRUE)
+        p2 = p1 + ggplot2::geom_line(data = dat, aes( y = .data$yvals, color = .data$varnames, linetype = .data$varnames), size = linesize, na.rm=TRUE)
       }
       if (plottype == 'Mixedplot')
       {
         #a mix of lines and points. for this, the dataframe needs to contain an extra column indicating line or point
-        p1a = p1 + ggplot2::geom_line(data = dplyr::filter(dat,style == 'line'), size = linesize)
-        p2 = p1a + ggplot2::geom_point(data = dplyr::filter(dat,style == 'point'), size = 2.5*linesize)
+        p1a = p1 + ggplot2::geom_line(data = dplyr::filter(dat,style == 'line'), aes( y = .data$yvals, color = .data$varnames, linetype = .data$varnames), size = linesize)
+        p2 = p1a + ggplot2::geom_point(data = dplyr::filter(dat,style == 'point'), aes( y = .data$yvals, color = .data$varnames, shape = .data$varnames), size = 2.5*linesize)
       }
 
 
-
-	 #set x-axis. no numbering/labels on x-axis for boxplots
-	 if (plottype == 'Boxplot')
-      {
-        p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax), breaks = NULL, labels = NULL)
-        p3 = p3 + ggplot2::labs(x = NULL)
-        }
+  	 #set x-axis. no numbering/labels on x-axis for boxplots
+  	 if (plottype == 'Boxplot')
+     {
+          p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax), breaks = NULL, labels = NULL)
+          p3 = p3 + ggplot2::labs(x = NULL)
+      }
       else
       {
-        p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax))
-        if (!is.null(resnow$xlab)) { p3 = p3 + ggplot2::xlab(resnow$xlab) }
+          p3 = p2 + ggplot2::scale_x_continuous(trans = xscaletrans, limits=c(xmin,xmax))
+          if (!is.null(resnow$xlab)) { p3 = p3 + ggplot2::xlab(resnow$xlab) }
       }
 
       #apply y-axis and if provided, label
@@ -169,6 +178,9 @@ generate_ggplot <- function(res)
       #modify overall theme
       p5 = p4 + ggplot2::theme_bw(base_size = 18)
 
+      #default palette is set, overwritten if user provided
+      plotpalette = cbfpalette
+      if (!is.null(resnow$palette)) {plotpalette = resnow$palette }
 
       #do legend if TRUE or not provided
       if (is.null(resnow$makelegend) || resnow$makelegend)
@@ -183,21 +195,41 @@ generate_ggplot <- function(res)
         }
         legendtitle = ifelse(is.null(resnow$legendtitle), "Variables", resnow$legendtitle)
 
-        p5a = p5 + ggplot2::theme(legend.key.width = grid::unit(3, "line"))
-        p5b = p5a + ggplot2::theme(legend.position = legendlocation)
-        p5c = p5b + ggplot2::scale_linetype_discrete(name = legendtitle) + ggplot2::scale_shape_discrete(name = legendtitle)
-        p5d = p5c + ggplot2::scale_colour_discrete(name = legendtitle)
-        p6 = p5d + ggplot2::guides(fill=guide_legend(title.position="top", nrow=3, byrow=TRUE))
-      }
+
+        if (plottype != 'Mixedplot')
+        {
+          nvars = length(unique(dat$varnames))
+          p5a = p5 + ggplot2::guides(col = ggplot2::guide_legend(nrow=2, byrow=TRUE,title.position = 'left'))
+          p5b = p5a + ggplot2::theme(legend.position = legendlocation) #default is top
+          p5c = p5b + ggplot2::theme(legend.key.width = grid::unit(3, "line")) #line thickness
+          p5d = p5c + ggplot2::scale_colour_manual(name = legendtitle, values=plotpalette[1:nvars]) #color for each variable
+          p5e = p5d + ggplot2::scale_linetype_discrete(name = legendtitle) #line type for each variable
+          pfinal = p5e + ggplot2::scale_shape_discrete(name = legendtitle) #symbol type for symbols
+        }
+        if (plottype == 'Mixedplot')
+        {
+          #trying to get legend right for combined line and symbol plots
+          #not fully working yet
+          #for data/symbols, legend still shows both lines and symbols, no matter what the plot is
+          # Compute the number of types and methods
+          npoints = length(unique(dplyr::filter(dat,style == 'point')$varnames))
+          nlines = length(unique(dplyr::filter(dat,style == 'line')$varnames))
+          p5a = p5 + ggplot2::guides(col = ggplot2::guide_legend(nrow=2, byrow=TRUE,title.position = 'left'))
+          p5b = p5a + ggplot2::theme(legend.position = legendlocation) #default is top
+          p5c = p5b + ggplot2::theme(legend.key.width = grid::unit(3, "line")) #line thickness
+          p5d = p5c + ggplot2::scale_colour_manual(name = legendtitle, values=plotpalette[1:(nlines+npoints)]) #color for each variable
+          p5e = p5d + ggplot2::scale_linetype_discrete(name = legendtitle, guide = FALSE) #symbol type for symbols; here is some trickery to make the legend look combined (turn off legend title/name)
+          pfinal = p5e + ggplot2::scale_shape_discrete(name = "", guide = FALSE) #symbol type for symbols
+        }
+      } #end doing legend
       else
       {
-          p6 = p5 + ggplot2::theme(legend.position="none")
+        pfinal = p5 + ggplot2::theme(legend.position="none") + ggplot2::scale_colour_manual(values=plotpalette)
       }
 
-      #modify overall theme
-      pfinal = p6
-      
       allplots[[n]] = pfinal
+
+
 
     } #end loop over individual plots