+++ tests/testthat/test-modelexploration.R
@@ -18,12 +18,14 @@ test_that("test that modelexploration app returns the proper plots",
 
             modelsettings$modeltype = '_modelexploration_'
             modelsettings$nplots = 1
-            modelsettings$simfunction = 'simulate_modelexploration'
+            modelsettings$simfunction = 'simulate_basicbacteria_modelexploration'
             modelsettings$plotscale = 'x'
 
             result = run_model(modelsettings)
 
-            testthat::expect_is( generate_ggplot(result), "ggplot" )
+            p = DSAIRM::generate_ggplot(result)
+            expect_true( is.ggplot(p))
+
             testthat::expect_is( generate_plotly(result), "plotly" )
             testthat::expect_is( generate_text(result), "html" )
             testthat::expect_is( generate_text(result), "character" )
@@ -32,9 +34,9 @@ test_that("test that modelexploration app returns the proper plots",
             modelsettings$pardist = "log"
             modelsettings$samplepar = "r"
             result = run_model(modelsettings)
-            testthat::expect_is( generate_ggplot(result), "ggplot" )
 
-
+            p = DSAIRM::generate_ggplot(result)
+            expect_true( is.ggplot(p))
 
             })