+++ tests/testthat/test-run_model.R
@@ -37,4 +37,28 @@ test_that("run_model correctly runs different models",
             result = run_model(modelsettings)
             testthat::expect_equal(result, "Model run failed. Maybe unreasonable parameter values?")
 
+            #no model type provided, should fail
+            modelsettings =  list(B = 10, I = 1, g = 1, Bmax = 1e6, dB = 0.1, k = 1e-07, r = 1e-3, dI = 1, tstart = 0, tfinal = 50, dt = 0.01, plotscale = 'x', nplots = 1)
+            modelsettings$simfunction = 'simulate_basicbacteria_discrete'
+            result = run_model(modelsettings)
+            testthat::expect_equal(result, "List element modeltype must be provided.")
+
+            #another model try
+            packagename = 'DSAIRM'
+
+            #find path to apps
+            appdir = system.file("appinformation", package = packagename) #find path to apps
+            #load app table that has all the app information
+            at = read.table(file = paste0(appdir,"/apptable.tsv"), sep = '\t', header = TRUE)
+            appName = "bacteriaexploration"
+            appsettings <- as.list(at[which(at$appid == appName),])
+            modelsettings =  list(B = 100, I = 10, g=2, Bmax=1e5, dB=1, k=1e-4, r=1e-4, dI=2, tstart = 0, tfinal = 300, dt = 0.1, samples = 10, parmin=2, parmax=10, samplepar='g',  pardist = 'lin')
+            modelsettings = c(appsettings,modelsettings)
+            result = run_model(modelsettings)
+            Bpeak = round(result[[1]]$dat$Bpeak[1])
+
+            testthat::expect_equal(Bpeak, 49505)
+
+
+
 })