+++ NEWS.md
@@ -1,3 +1,19 @@
+# DSAIRM 0.8.6
+
+* CRAN release
+* added another app
+* fixes and updates in instructions
+
+# DSAIRM 0.8.5
+
+* updated some internals to better style
+* updated plotting functions
+
+# DSAIRM 0.8.3
+
+* updated fitting apps
+
+
 # DSAIRM 0.8.2
 
 * small bug fixes and feature enhancements